Refactored and upgraded graph for prod
