+++ src/components/Graph.module.scss
