+++ src/pages/Home.tsx
@@ -1,5 +1,4 @@
-//import SigmaGraph from "../components/Graph";
-import DemoGraph from "../components/DemoGraph";
+import SigmaGraph from "../components/Graph";
 
 function Home() {
     return (
@@ -7,7 +6,7 @@ function Home() {
             <h1>Hollywood Director Networks</h1>
             <p>This is an exploration of how directors in Hollywood use similar crew members throughout multiple productions.</p>
             <br/>
-            <DemoGraph/>
+            <SigmaGraph/>
         </div>
     );
 }
